Project architecture changed to modular; add BaseSectionTitle ui component; pagination deleted from ListTracks
