+++ src/components/Navbar.vue
@@ -1,4 +1,4 @@
-<script setup lang="ts">
+<script setup>
 import { RouterLink } from 'vue-router';
 </script>
