+++ src/main.js
@@ -2,7 +2,7 @@ import { createApp } from 'vue'
 import { createPinia } from 'pinia'
 import App from './App.vue'
 import FontAwesomeIcon from "./plugins/fontawesome"
-import './assets/styles/main.css'
+import './shared/styles/main.css'
 
 import 'bootstrap/dist/css/bootstrap.min.css'
 import 'bootstrap'
