+++ src/modules/player/PlayerVolumeSlider.vue
