+++ src/modules/tracks/UploadTrackForm.vue
@@ -1,8 +1,8 @@
 <script setup>
 import {ref} from "vue";
-import BaseButton from "@/components/base/BaseButton.vue";
-import BaseInput from "@/components/base/BaseInput.vue";
-import BaseFileInput from "@/components/base/BaseFileInput.vue";
+import BaseButton from "@/ui/BaseButton.vue";
+import BaseInput from "@/ui/BaseInput.vue";
+import BaseFileInput from "@/ui/BaseFileInput.vue";
 
 const musicFile = ref(null);
 const coverFile = ref(null);
