+++ src/router/index.js
@@ -1,10 +1,10 @@
 import { createWebHistory, createRouter } from 'vue-router'
 
-import AllTracks from "@/views/AllTracks.vue";
-import FavoriteTracks from "@/views/FavoriteTracks.vue";
-import UploadTracks from "@/views/UploadTracks.vue";
-import Login from '@/views/Login.vue';
-import Register from '@/views/Register.vue';
+import AllTracks from "@/pages/AllTracks.vue";
+import FavoriteTracks from "@/pages/FavoriteTracks.vue";
+import UploadTracks from "@/pages/UploadTracks.vue";
+import Login from '@/pages/Login.vue';
+import Register from '@/pages/Register.vue';
 
 const routes = [
     { path: '/', component: AllTracks },
