+++ src/shared/helpers/useIsDesktop.js
